Prediction Object for Forecasting

This object wraps the predictions returned by a learner of class LearnerForecast, i.e. the predicted response and standard error.

Format

R6::R6Class object inheriting from Prediction.

Construction

p = PredictionForecast$new(task = NULL, row_ids = task$row_ids, truth = task$truth(), response = NULL, se = NULL)

task :: TaskRegrForecast
Task, used to extract defaults for row_ids and truth.

row_ids :: integer()
Row ids of the predicted observations, i.e. the row ids of the test set.

truth :: numeric()
True (observed) outcome.

response :: numeric()
Object of numeric response values that can be coerced to a data.table.

se :: numeric()
Object of numeric standard erros that can be coerced to a data.table.

Fields

All fields from Prediction, and additionally:

response :: numeric()
Access to the stored predicted response.

se :: numeric()
Access to the stored standard error.

The field task_type is set to "forecast".

Methods

All Methods from Prediction, and additionally:

conf_int(level)
Access to the stored predicted response.

Examples

task = mlr3::tsk("airpassengers")
learner=LearnerRegrForecastAutoArima$new()
learner$train(task, 1:30)
p=learner$predict(task, 31:50)