Forecast Task

This task specializes Task and TaskSupervised for forecasting problems. The target column is assumed to be numeric. The task_type is set to "forecast" `.

Format

R6::R6Class object inheriting from Task/TaskSupervised.

Construction

t = TaskRegrForecast$new(id, backend, target, date_col)

id :: character(1)
Identifier for the task.

backend :: DataBackend
Either a DataBackendLong, a object of class ts or a data.frame with specified date column or any object which is convertible to a DataBackend with as_data_backend(). E.g., a object of class dts will be converted to a DataBackendLong.

target :: character(n)
Name of the target column(s).

Fields

All methods from TaskSupervised, and additionally:

date_col :: character(1)
Name of the date column.

Methods

See TaskSupervised, additionally:

date(row_ids = NULL) :: data.table
(integer() | character()) -> named list()
Returns the date column.

See also