Forecast Holdout Resampling

Splits data into a training set and a test set. Parameter ratio determines the ratio of observation going into the training set (default: 2/3).

Format

R6::R6Class inheriting from Resampling.

Construction

ResamplingForecastHoldout$new()
mlr_resamplings$get("forecast.holdout")
rsmp("forecast.holdout")

Fields

See Resampling.

Methods

See Resampling.

Parameters

ratio :: numeric(1)
Ratio of observations to put into the training set.

See also

Examples

 #Create a task with 10 observations
task = mlr3::tsk("airpassengers")
task$filter(1:10)

#Instantiate Resampling
rfho = mlr3::rsmp("forecastHoldout", ratio = 0.5)
rfho$instantiate(task)

# Individual sets:
rfho$train_set(1)

#> [1] 1 2 3 4 5

rfho$test_set(1)

#> [1]  6  7  8  9 10

intersect(rfho$train_set(1), rfho$test_set(1))

#> integer(0)

# Internal storage:
rfho$instance # simple list

#> $train
#> [1] 1 2 3 4 5
#> 
#> $test
#> [1]  6  7  8  9 10
#>